% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onyx.R
\name{onyx}
\alias{onyx}
\title{onyx}
\usage{
onyx(model = NULL, onyxfile = NULL, batch = NULL)
}
\arguments{
\item{model}{An OpenMx or lavaan model. Defaults to NULL. If omitted, Onyx is started with an empty desktop, otherwise
the specified model is rendered as path diagram}

\item{onyyfile}{path to Onyx executable (onyx-***.jar). Defaults to NULL. If NULL, Onyx searches local directors, otherwise downloads Onyx from official repository.}

\item{mode}{batch operation mode (experimental)}
}
\description{
Starts Onyx executable to display a path diagram from either an OpenMx or
lavaan model representation.
}
\examples{

require(OpenMx)
data(demoOneFactor)
manifests <- names(demoOneFactor)
latents <- c("G")
factorModel <- mxModel("One Factor",
                       type="RAM",
                       manifestVars = manifests,
                       latentVars = latents,
                       mxPath(from=latents, to=manifests),
                       mxPath(from=manifests, arrows=2),
                       mxPath(from=latents, arrows=2,
                              free=FALSE, values=1.0),
                       mxData(cov(demoOneFactor), type="cov",
                              numObs=500))
fit <- mxRun(factorModel)
\dontrun{
onyx(fit)
}

}
